be sure to use system perl for fixups
It’s important to avoid picking up a perl from $PATH because that points
to the plenv shim directory and the plenv shims add the selected perl to
the $PATH before invoking the binary in question. So the fixup does not
get to see the pristine $PATH and ends up etching the temporary addition
into $PATH permanently (for the remainder of the session).

Switching to the system perl defeats the Pod::Perldoc::ToTerm detection,
but that formatter is now the default. Passing the -R switch to less via
PERLDOC_PAGER is needed until Pod::Perldoc 3.25 from perl 5.22 – but can
be done unconditionally.